requirements.txt
pandas
scikit-learn
joblib
streamlit
requests
langchain
langchain-huggingface
python-dotenv
matplotlib
